Add files via upload
So here's the deal, I got the API spitting out data, but I have no way to turn it to a json to navigate. jsontest.py is me trying to test json functionality, mean.py is the meat and potatoes of the backend. the screenshot is the one I uploaded to discord. stuffs.json is me trying to manually make a json file (probably badly) and the Weather Testing API files is me trying to run my code, and for those lacking in jupyter notebooks, I added the html so you can take a look along. All in all, I wish I could just easy call the api for a json, this makes me want to die.